+++ .env
@@ -8,4 +8,4 @@
 # flags must begin with flag_ and are case sensitive
 
 flag_google_import=true
-flag_location_refactor=true
+flag_better_location_status_checks=true
